+++ trainer/src/nn.rs
@@ -12,7 +12,7 @@ pub const OBSERVATION_ZERO: Observation = [0.; STATE_SIZE];
 
 pub const BATCH_SIZE: usize = 256;
 pub const EPOCHS: usize = 50;
-pub const LEARNING_RATE: f32 = 0.0025;
+pub const LEARNING_RATE: f64 = 0.0025;
 // pub const DECAY: f32 = 0.001;
 
 pub const HIDDEN_SIZE: usize = 16;
